+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - fix: Update README #1176
 - fix: “Close Tab” icon too small #1156
 - new: switched from deepl to fairtranslate
+- new: edit custom redirects
 
 ### v 16 "Alexander Berkmann"
 - new: tracking-URL-popup